requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
llama-cpp-python==0.2.11